Authentication/pyproject.toml
[tool.poetry]
name = "authentication"
version = "0.1.0"
description = ""
authors = ["Berna RQ"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
motor = "^3.6.0"
logging = "^0.4.9.6"
click = "^8.1.8"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
